Refactor command creation
Now each command's definition and handler goes in it's own file and is added to the bot in bot.New()
